# static/openapi.yaml
openapi: 3.1.1
info:
  title: Tiny URL
  description: A very much NOT prodution ready URL shortener
  version: 1.0.0
paths:
  /{id}:
    get:
      summary: Translate ID into the original URL and redirects request to it.
      parameters:
        - name: id
          in: path 
          description: ID of the tiny URL
          required: true
          schema:
            type: string
      responses:
        "302":
          description: URL found
          headers:
            Location:
              description: Original URL
              required: true
              style: simple
              schema:
                type: string
              example: https://example.com
        "404":
          description: URL not found
  /api/urls:
    post:
      summary: Create a new tiny urls
      description: Create a new tiny urls; if the field "from" is not provided a random ID will be generated by the server 
      requestBody:
        description: Redirection entry
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TinyUrlMapping"
      responses:
        "200":
          description: Tiny URL created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TinyUrlMapping"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/urls/{id}:
    get:
      summary: Fetches the a tiny url mapping
      parameters:
        - name: id
          in: path
          description: ID of the tiny URL
          required: true
          style: simple
          schema:
            type: string
            example: coolid007
      responses:
        "200":
          description: Tiny URL mapping
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TinyUrlMapping"
        "404":
          description: Mapping not found
    delete:
      summary: Deletes a tiny url mapping
      parameters:
        - name: id
          in: path
          description: ID of the tiny URL
          required: true
          style: simple
          schema:
            type: string
            example: coolid007
      responses:
        "200":
          description: Tiny URL mapping deleted
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
components: 
  schemas:
    TinyUrlMapping:
      type: object
      required:
        - to
      properties:
        from:
          description: ID to use on the tiny URL
          type: string
          example: coolid007
        to:
          description: Original URL
          type: string
          example: https://example.com
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Description of the error